further extracted response helper so I can use it with docdata also
