+++ api/Template/ListVersions.php
@@ -5,6 +5,7 @@ namespace HomeCEU\DTS\Api\Template;
 
 
 use Exception;
+use HomeCEU\DTS\Api\ResponseHelper;
 use HomeCEU\DTS\Entity\Template;
 use HomeCEU\DTS\Persistence\CompiledTemplatePersistence;
 use HomeCEU\DTS\Persistence\TemplatePersistence;
@@ -33,14 +34,14 @@ class ListVersions {
       $responseData = [
           'total' => count($templates),
           'items' => array_map(function (Template $t) {
-            return TemplateResponseHelper::templateDetailModel($t);
+            return ResponseHelper::templateDetailModel($t);
           },$templates)
       ];
       return $response
           ->withStatus(200)
           ->withJson($responseData);
     } catch (Exception $e) {
-      return $response->withStatus(500, "failed to list templates");
+      return $response->withStatus(500, __CLASS__." failure");
     }
   }
 }
